Inline style experiment (#48)
* working ssr
* remove logs
